# docs/indexer/indexer_swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Meta File System Indexer Service API, provides file query and download functionality",
        "title": "Meta File System Indexer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7281",
    "basePath": "/api/v1",
    "paths": {
        "/avatars": {
            "get": {
                "description": "Query avatar list with cursor pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer Avatar Query"
                ],
                "summary": "Query avatar list",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Cursor (last avatar ID)",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/meta-file-system_controller_respond.IndexerAvatarListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/avatars/accelerate/address/{address}": {
            "get": {
                "description": "Redirect to OSS URL for latest avatar by address, supports preview/thumbnail processing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer Avatar Query"
                ],
                "summary": "Get accelerated avatar by address (redirect to OSS)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Process type: preview (640px), thumbnail (128x128), empty for original",
                        "name": "process",
                        "in": "query"
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Redirect to OSS URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/avatars/accelerate/content/{pinId}": {
            "get": {
                "description": "Redirect to OSS URL for avatar content by PIN ID, supports preview/thumbnail processing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer Avatar Query"
                ],
                "summary": "Get accelerated avatar content (redirect to OSS)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PIN ID",
                        "name": "pinId",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Process type: preview (640px), thumbnail (128x128), empty for original",
                        "name": "process",
                        "in": "query"
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Redirect to OSS URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/avatars/accelerate/metaid/{metaId}": {
            "get": {
                "description": "Redirect to OSS URL for latest avatar by MetaID, supports preview/thumbnail processing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer Avatar Query"
                ],
                "summary": "Get accelerated avatar by MetaID (redirect to OSS)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MetaID",
                        "name": "metaId",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Process type: preview (640px), thumbnail (128x128), empty for original",
                        "name": "process",
                        "in": "query"
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Redirect to OSS URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/avatars/address/{address}": {
            "get": {
                "description": "Query the latest avatar information by address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer Avatar Query"
                ],
                "summary": "Get latest avatar by address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/meta-file-system_controller_respond.IndexerAvatarResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/avatars/content/{pinId}": {
            "get": {
                "description": "Get avatar content by PIN ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Indexer Avatar Query"
                ],
                "summary": "Get avatar content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PIN ID",
                        "name": "pinId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/avatars/metaid/{metaId}": {
            "get": {
                "description": "Query the latest avatar information by MetaID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer Avatar Query"
                ],
                "summary": "Get latest avatar by MetaID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MetaID",
                        "name": "metaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/meta-file-system_controller_respond.IndexerAvatarResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/files": {
            "get": {
                "description": "Query file list with cursor pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer File Query"
                ],
                "summary": "Query file list",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Cursor",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/meta-file-system_controller_respond.IndexerFileListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                        }
                    }
                }
            }
        },
...
        "/files/content/{pinId}": {
            "get": {
                "description": "Get file content by PIN ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Indexer File Query"
                ],
                "summary": "Get file content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PIN ID",
                        "name": "pinId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/meta-file-system_controller_respond.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "meta-file-system_controller_respond.IndexerAvatarListResponse": {
            "type": "object",
            "properties": {
                "avatars": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meta-file-system_controller_respond.IndexerAvatarResponse"
                    }
                },
                "has_more": {
                    "type": "boolean",
                    "example": true
                },
                "next_cursor": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "meta-file-system_controller_respond.Response": {
            "description": "Unified API response structure",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "processingTime": {
                    "type": "integer",
                    "example": 123
                }
            }
        }
    }
}